fix: q1 cast each num to float

+++ q1.sql
@@ -49,8 +49,8 @@ CREATE VIEW TotalPercentageSold AS
     SELECT 
         t.concert_id,
         (CASE WHEN s.sold_count IS NULL THEN 0
-        ELSE (s.sold_count / t.ticket_count) * 100
-        END)::REAL AS "percentage_sold"
+        ELSE (s.sold_count::FLOAT / t.ticket_count::FLOAT) * 100
+        END) AS "percentage_sold"
     FROM TotalTicketCount t
         LEFT JOIN TotalTicketSoldCount s ON s.concert_id = t.concert_id;
